+++ backend/keystone.ts
@@ -1,30 +1,30 @@
-import { config } from '@keystone-next/keystone/schema';
-import { statelessSessions } from '@keystone-next/keystone/session';
-import { createAuth } from '@keystone-next/auth';
+import { config } from "@keystone-next/keystone/schema";
+import { statelessSessions } from "@keystone-next/keystone/session";
+import { createAuth } from "@keystone-next/auth";
 
-import { lists } from './schema';
+import { lists } from "./schemas";
 
 let sessionSecret = process.env.SESSION_SECRET;
 
 if (!sessionSecret) {
-  if (process.env.NODE_ENV === 'production') {
+  if (process.env.NODE_ENV === "production") {
     throw new Error(
-      'The SESSION_SECRET environment variable must be set in production'
+      "The SESSION_SECRET environment variable must be set in production"
     );
   } else {
-    sessionSecret = '-- DEV COOKIE SECRET; CHANGE ME --';
+    sessionSecret = "-- DEV COOKIE SECRET; CHANGE ME --";
   }
 }
 
 let sessionMaxAge = 60 * 60 * 24 * 30; // 30 days
 
 const { withAuth } = createAuth({
-  listKey: 'User',
-  identityField: 'email',
-  secretField: 'password',
-  sessionData: 'name',
+  listKey: "User",
+  identityField: "email",
+  secretField: "password",
+  sessionData: "name",
   initFirstItem: {
-    fields: ['name', 'email', 'password'],
+    fields: ["name", "email", "password"],
   },
 });
 
@@ -36,8 +36,10 @@ const session = statelessSessions({
 export default withAuth(
   config({
     db: {
-      adapter: 'prisma_postgresql',
-      url: process.env.DATABASE_URL || 'postgres://postgres:postgres@localhost:5432/ecommerce?schema=public',
+      adapter: "prisma_postgresql",
+      url:
+        process.env.DATABASE_URL ||
+        "postgres://postgres:postgres@localhost:5432/ecommerce?schema=public",
     },
     ui: {
       isAccessAllowed: (context) => !!context.session?.data,
