+++ backend/schemas/index.ts
@@ -1,0 +1,6 @@
+import { createSchema, list } from "@keystone-next/keystone/schema";
+import { User } from "./User";
+
+export const lists = createSchema({
+  User,
+});
